Move SpecialFolderSelection out of the account object
