+++ legacy/account/src/main/java/app/k9mail/legacy/account/SpecialFolderSelection.kt
@@ -1,0 +1,6 @@
+package app.k9mail.legacy.account
+
+enum class SpecialFolderSelection {
+    AUTOMATIC,
+    MANUAL,
+}
