+++ legacy/core/src/main/java/com/fsck/k9/AccountPreferenceSerializer.kt
@@ -4,7 +4,6 @@ import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
 import app.k9mail.legacy.account.Account.ShowPictures
-import app.k9mail.legacy.account.Account.SpecialFolderSelection
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_MAXIMUM_AUTO_DOWNLOAD_MESSAGE_SIZE
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_MESSAGE_FORMAT
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_MESSAGE_FORMAT_AUTO
@@ -24,6 +23,7 @@ import app.k9mail.legacy.account.Expunge
 import app.k9mail.legacy.account.FolderMode
 import app.k9mail.legacy.account.Identity
 import app.k9mail.legacy.account.SortType
+import app.k9mail.legacy.account.SpecialFolderSelection
 import app.k9mail.legacy.notification.NotificationLight
 import app.k9mail.legacy.notification.NotificationSettings
 import app.k9mail.legacy.notification.NotificationVibration
