+++ legacy/core/src/main/java/com/fsck/k9/preferences/AccountSettingsDescriptions.java
@@ -19,7 +19,7 @@ import app.k9mail.legacy.account.Account.MessageFormat;
 import app.k9mail.legacy.account.Account.QuoteStyle;
 import app.k9mail.legacy.account.Account.ShowPictures;
 import app.k9mail.legacy.account.SortType;
-import app.k9mail.legacy.account.Account.SpecialFolderSelection;
+import app.k9mail.legacy.account.SpecialFolderSelection;
 import app.k9mail.legacy.di.DI;
 import com.fsck.k9.K9;
 import com.fsck.k9.core.R;
